+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 
-members = [ "dpc", "algebra", "r1cs-core", "r1cs-lib", "gm17", "snark-gadgets", "bench-utils" ]
+members = [ "dpc", "algebra", "r1cs-core", "r1cs-std", "gm17", "bench-utils" ]
 
 [profile.release]
 opt-level = 3
