+++ r1cs-std/src/fields/fp.rs
@@ -9,10 +9,7 @@ use crate::{
     uint8::UInt8,
 };
 
-use crate::utils::{
-    AllocGadget, CondSelectGadget, ConditionalEqGadget, EqGadget, NEqGadget, ToBitsGadget,
-    ToBytesGadget, TwoBitLookupGadget,
-};
+use crate::prelude::*;
 
 use crate::{
     Assignment,
