+++ r1cs-std/src/fields/fp2.rs
@@ -1,21 +1,13 @@
 use algebra::{
-    fields::{Field, Fp2, Fp2Parameters},
-    PrimeField,
+    fields::{Fp2, Fp2Parameters},
+    Field, PrimeField,
 };
 use r1cs_core::{ConstraintSystem, SynthesisError};
 use std::{borrow::Borrow, marker::PhantomData};
 
-use super::{fp::FpGadget, FieldGadget};
-
-use crate::{
-    boolean::Boolean,
-    uint8::UInt8,
-    utils::{
-        AllocGadget, CondSelectGadget, ConditionalEqGadget, EqGadget, NEqGadget, ToBitsGadget,
-        ToBytesGadget, TwoBitLookupGadget,
-    },
-    Assignment, ConstraintVar,
-};
+use crate::fields::fp::FpGadget;
+use crate::prelude::*;
+use crate::{Assignment, ConstraintVar};
 
 #[derive(Derivative)]
 #[derivative(Debug(bound = "P: Fp2Parameters, ConstraintF: PrimeField"))]
