--- snark-gadgets/src/bits/mod.rs
@@ -1,3 +1,0 @@
-pub mod boolean;
-pub mod uint32;
-pub mod uint8;
